+++ example/lib/data/mock_universal_ble.dart
@@ -20,7 +20,7 @@ class MockUniversalBle extends UniversalBlePlatform {
       CharacteristicProperty.read,
       CharacteristicProperty.write,
       CharacteristicProperty.notify,
-    ]),
+    ], []),
   ]);
 
   @override
@@ -53,7 +53,8 @@ class MockUniversalBle extends UniversalBlePlatform {
   }
 
   @override
-  Future<List<BleService>> discoverServices(String deviceId) async {
+  Future<List<BleService>> discoverServices(
+      String deviceId, bool withDescriptors) async {
     return [_mockService];
   }