+++ lib/src/universal_ble_platform_interface.dart
@@ -56,7 +56,8 @@ abstract class UniversalBlePlatform {
 
   Future<void> disconnect(String deviceId);
 
-  Future<List<BleService>> discoverServices(String deviceId);
+  Future<List<BleService>> discoverServices(
+      String deviceId, bool withDescriptors);
 
   Future<void> setNotifiable(String deviceId, String service,
       String characteristic, BleInputProperty bleInputProperty);